% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Partitioning.R
\name{came}
\alias{came}
\title{Computes a partitioning for an \code{FME}}
\usage{
came(
  effects,
  number.partitions = NULL,
  max.cov = Inf,
  rp.method = "ctree",
  tree.control = NULL
)
}
\arguments{
\item{effects}{an \code{FME} object with FMEs computed.}

\item{number.partitions}{the exact number of partitions required.
Either \code{number.partitions} or \code{max.cov} can be specified.}

\item{max.cov}{the maximum coefficient of variation required in each partition.
Among multiple partitionings with this criterion identified, the one with lowest number of partitions is selected.
Either \code{number.partitions} or \code{max.cov} can be specified.}

\item{rp.method}{one of \code{"ctree"} or \code{"rpart"}. The RP algorithm used for growing the decision tree.}

\item{tree.control}{control parameters for the RP algorithm. One of \code{"ctree.control(...)"} or \code{"rpart.control(...)"}.
#' @return \code{Partitioning} object with identified feature subspaces.}
}
\description{
This is a wrapper function that creates the correct subclass of \code{Partitioning}.
It computes feature subspaces for semi-global interpretations of FMEs via recursive partitioning (RP).
}
\examples{
# Train a model and compute FMEs:
data("Boston", package = "MASS")
forest = randomForest(medv ~ ., data = Boston)
effects = fme(model = forest, data = Boston, target = "medv", feature = "rm",
              step.size = 1, ep.method = "envelope", )

# Find a partitioning with exactly 3 subspaces:
subspaces = came(effects, number.partitions = 3)

# Find a partitioning with a maximum coefficient of variation of 4, use `rpart`:
subspaces = came(effects, max.cov = 4, rp.method = "rpart")

# Analyze results:
summary(subspaces)
plot(subspaces)

# Extract results:
subspaces$results
subspaces$tree

}
\references{
Scholbeck, C. A., Casalicchio, G., Molnar, C., Bischl, B., & Heumann, C. (2022). Marginal Effects for Non-Linear Prediction Functions.
}
