% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FME.R
\name{fme}
\alias{fme}
\title{User-friendly function to compute FMEs}
\usage{
fme(
  model,
  data,
  target,
  feature,
  step.size,
  ep.method = "none",
  compute.nlm = TRUE,
  nlm.intervals = 1
)
}
\arguments{
\item{model}{\code{Predictor} object.}

\item{data}{data.table or data.frame object.}
}
\description{
This is a wrapper function that provides a user-friendly interface by abstracting away the R6 functionality of the package.
}
\examples{
R code here
...
}
