% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{fmeffects-package}
\alias{fmeffects}
\alias{fmeffects-package}
\title{fmeffects}
\description{
Computes forward marginal effects (FME) for arbitrary supervised machine learning models. You provide a model and data, and 'fmeffects' gives you feature effects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/holgstr/fmeffects}
  \item Report bugs at \url{https://github.com/holgstr/fmeffects/issues}
}

}
\author{
\strong{Maintainer}: Holger Löwe \email{hbj.loewe@gmail.com}

Authors:
\itemize{
  \item Christian Scholbeck \email{christian.scholbeck@stat.uni-muenchen.de}
}

Other contributors:
\itemize{
  \item Christian Heumann \email{christian.heumann@stat.uni-muenchen.de} [reviewer]
  \item Bernd Bischl \email{bernd.bischl@stat.uni-muenchen.de} [reviewer]
  \item Giuseppe Casalicchio \email{giuseppe.casalicchio@stat.uni-muenchen.de} [reviewer]
}

}
