% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartitioningCtree.R
\name{PartitioningCtree}
\alias{PartitioningCtree}
\title{PartitioningCtree}
\description{
This task specializes \link{Partitioning} for the \code{ctree} algorithm for recursive partitioning.

It is recommended to use \code{\link[=came]{came()}} for construction of Partitioning objects.
}
\section{Super class}{
\code{\link[fmeffects:Partitioning]{fmeffects::Partitioning}} -> \code{PartitioningCtree}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PartitioningCtree-new}{\code{PartitioningCtree$new()}}
\item \href{#method-PartitioningCtree-clone}{\code{PartitioningCtree$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="fmeffects" data-topic="Partitioning" data-id="compute"><a href='../../fmeffects/html/Partitioning.html#method-Partitioning-compute'><code>fmeffects::Partitioning$compute()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fmeffects" data-topic="Partitioning" data-id="plot"><a href='../../fmeffects/html/Partitioning.html#method-Partitioning-plot'><code>fmeffects::Partitioning$plot()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartitioningCtree-new"></a>}}
\if{latex}{\out{\hypertarget{method-PartitioningCtree-new}{}}}
\subsection{Method \code{new()}}{
Create a new PartitioningCtree object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartitioningCtree$new(object, method, value, tree.control = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object}}{an \code{FME} object with results computed.}

\item{\code{method}}{the method for finding feature subspaces.}

\item{\code{value}}{the value of \code{method}.}

\item{\code{tree.control}}{control parameters for the RP algorithm.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartitioningCtree-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PartitioningCtree-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartitioningCtree$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
