% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{plot.ForwardMarginalEffect}
\alias{plot.ForwardMarginalEffect}
\title{Plots an ForwardMarginalEffect object.}
\usage{
\method{plot}{ForwardMarginalEffect}(fme, with.nlm = FALSE, jitter = c(0, 0))
}
\arguments{
\item{fme}{object of class \code{ForwardMarginalEffect}.}

\item{with.nlm}{plots NLMs alongside FMEs, defaults to \code{FALSE}.}

\item{jitter}{Jitters data. A two-dimensional numeric vector, corresponds to \code{"width"} and \code{"height"}. See \code{?ggplot2::geom_jitter} for details.
Not available if \code{step.type} is categorical.
Defaults to no jittering, i.e., c(0, 0).}
}
\description{
Plots an ForwardMarginalEffect object.
}
