% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{plot.Partitioning}
\alias{plot.Partitioning}
\title{Plots an FME Partitioning.}
\usage{
\method{plot}{Partitioning}(x, ...)
}
\arguments{
\item{x}{object of class \code{Partitioning}.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Plots an FME Partitioning.
}
