% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictorRandomForest.R
\name{PredictorRandomForest}
\alias{PredictorRandomForest}
\title{PredictorRandomForest}
\description{
This task specializes \link{Predictor} for \code{randomForest} regression models.
The \code{model} is assumed to be a \code{"randomForest"}.

It is recommended to use \code{\link[=makePredictor]{makePredictor()}} for construction of Predictor objects.
}
\section{Super class}{
\code{\link[fme:Predictor]{fme::Predictor}} -> \code{PredictorRandomForest}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PredictorRandomForest-new}{\code{PredictorRandomForest$new()}}
\item \href{#method-PredictorRandomForest-predict}{\code{PredictorRandomForest$predict()}}
\item \href{#method-PredictorRandomForest-clone}{\code{PredictorRandomForest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictorRandomForest-new"></a>}}
\if{latex}{\out{\hypertarget{method-PredictorRandomForest-new}{}}}
\subsection{Method \code{new()}}{
Create a new PredictorRandomForest object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictorRandomForest$new(model, data, target)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{\code{randomForest} object.}

\item{\code{data}}{The data used for computing FMEs, must be data.frame or data.table.}

\item{\code{target}}{A string specifying the target variable.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictorRandomForest-predict"></a>}}
\if{latex}{\out{\hypertarget{method-PredictorRandomForest-predict}{}}}
\subsection{Method \code{predict()}}{
Predicts on an observation \code{"newdata"}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictorRandomForest$predict(newdata)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newdata}}{the feature vector for which the target should be predicted.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictorRandomForest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PredictorRandomForest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictorRandomForest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
