% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{print.ForwardMarginalEffect}
\alias{print.ForwardMarginalEffect}
\title{Prints an ForwardMarginalEffect object.}
\usage{
\method{print}{ForwardMarginalEffect}(fme)
}
\arguments{
\item{fme}{object of class \code{ForwardMarginalEffect}.}
}
\description{
Prints an ForwardMarginalEffect object.
}
