% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{print.Partitioning}
\alias{print.Partitioning}
\title{Prints an FME Partitioning.}
\usage{
\method{print}{Partitioning}(partitioning)
}
\arguments{
\item{partitioning}{object of class \code{Partitioning}.}
}
\description{
Prints an FME Partitioning.
}
