% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{summary.AverageMarginalEffects}
\alias{summary.AverageMarginalEffects}
\title{Prints summary of an AverageMarginalEffects object.}
\usage{
\method{summary}{AverageMarginalEffects}(object, ...)
}
\arguments{
\item{object}{object of class \code{AverageMarginalEffects}.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Prints summary of an AverageMarginalEffects object.
}
