% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{summary.FME}
\alias{summary.FME}
\title{Prints summary of an FME object.}
\usage{
\method{summary}{FME}(fme)
}
\arguments{
\item{fme}{object of class \code{FME}.}
}
\description{
Prints summary of an FME object.
}
