% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{summary.ForwardMarginalEffect}
\alias{summary.ForwardMarginalEffect}
\title{Prints summary of an ForwardMarginalEffect object.}
\usage{
\method{summary}{ForwardMarginalEffect}(fme)
}
\arguments{
\item{fme}{object of class \code{ForwardMarginalEffect}.}
}
\description{
Prints summary of an ForwardMarginalEffect object.
}
