% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{plot.FME}
\alias{plot.FME}
\title{Plots an FME object.}
\usage{
\method{plot}{FME}(fme, with.nlm = FALSE)
}
\arguments{
\item{fme}{object of class \code{FME}.}

\item{with.nlm}{plots NLMs alongside FMEs, defaults to \code{FALSE}.}
}
\description{
Plots an FME object.
}
