% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{plot.ForwardMarginalEffect}
\alias{plot.ForwardMarginalEffect}
\title{Plots an ForwardMarginalEffect object.}
\usage{
\method{plot}{ForwardMarginalEffect}(x, ...)
}
\arguments{
\item{x}{object of class \code{ForwardMarginalEffect}.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Plots an ForwardMarginalEffect object.
}
