% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{print.ForwardMarginalEffect}
\alias{print.ForwardMarginalEffect}
\title{Prints an ForwardMarginalEffect object.}
\usage{
\method{print}{ForwardMarginalEffect}(x, ...)
}
\arguments{
\item{x}{object of class \code{ForwardMarginalEffect}.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Prints an ForwardMarginalEffect object.
}
