% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{print.Partitioning}
\alias{print.Partitioning}
\title{Prints an FME Partitioning.}
\usage{
\method{print}{Partitioning}(x, ...)
}
\arguments{
\item{x}{object of class \code{Partitioning}.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Prints an FME Partitioning.
}
