% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{summary.Partitioning}
\alias{summary.Partitioning}
\title{Prints summary of an FME Partitioning.}
\usage{
\method{summary}{Partitioning}(object, ...)
}
\arguments{
\item{object}{object of class \code{Partitioning}.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Prints summary of an FME Partitioning.
}
